Fix:  State machine cannot be killed with Ctrl+C
As the smach state machine was running on a separate thread the SIGINT signal was not forwarded to abort the task and the request_preempt method failed because the task server already terminated. Therefore the stop_task service could not be called anymore. This should now be fixed and additionally I also added a service call through which the state machine could be killed nicely (just as before) using the request_shutdown flag.